refactor(config): Regroup config related function in config module
To enforce module boundaries, we need to move some config related
function in the `core.config` module.

refs #521
